Module oid::prelude

Convenience module for quickly importing the public interface (e.g., use oid::prelude::*)

Re-exports

pub use crate::ObjectIdentifier;

pub use crate::ObjectIdentifierError;

Traits

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.